Tests: irrigations, dayplans

+++ watergenius/api/models/plans.py
@@ -7,5 +7,5 @@ class DayPlan(models.Model):
     dayplan_id = models.AutoField(primary_key=True)
     dayplan_gen_time = models.DateTimeField(auto_now=True)
     dayplan_time = models.DateTimeField()
-    dayplan_water_qtd = models.IntegerField()
+    dayplan_water_qtd = models.FloatField()
     dayplan_sub = models.ForeignKey(SubSpace, related_name='belongs_to_subspace', on_delete=models.CASCADE)
